+++ pkg/client/cli/util/connector.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"time"
 
-	"github.com/spf13/cobra"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	grpcCodes "google.golang.org/grpc/codes"
@@ -17,7 +16,6 @@ import (
 
 	"github.com/telepresenceio/telepresence/rpc/v2/connector"
 	"github.com/telepresenceio/telepresence/v2/pkg/client"
-	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/ann"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/output"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/errcat"
 	"github.com/telepresenceio/telepresence/v2/pkg/proc"
@@ -39,86 +37,6 @@ func GetConnectRequest(ctx context.Context) *connector.ConnectRequest {
 	if cr, ok := ctx.Value(connectRequest{}).(*connector.ConnectRequest); ok {
 		return cr
 	}
-	return nil
-}
-
-type userDaemonKey struct{}
-
-func GetUserDaemon(ctx context.Context) *UserDaemon {
-	if ud, ok := ctx.Value(userDaemonKey{}).(*UserDaemon); ok {
-		return ud
-	}
-	return nil
-}
-
-type sessionKey struct{}
-
-func GetSession(ctx context.Context) *Session {
-	if s, ok := ctx.Value(sessionKey{}).(*Session); ok {
-		return s
-	}
-	return nil
-}
-
-type cmdInitKey struct{}
-
-func WithCommandInitializer(ctx context.Context, cmdInit func(cmd *cobra.Command) error) context.Context {
-	return context.WithValue(ctx, cmdInitKey{}, cmdInit)
-}
-
-func InitCommand(cmd *cobra.Command) (err error) {
-	cmdInit, ok := cmd.Context().Value(cmdInitKey{}).(func(cmd *cobra.Command) error)
-	if !ok {
-		panic("no registered command initializer")
-	}
-	return cmdInit(cmd)
-}
-
-func CommandInitializer(cmd *cobra.Command) (err error) {
-	ctx := cmd.Context()
-	as := cmd.Annotations
-
-	if v, ok := as[ann.Session]; ok {
-		as[ann.UserDaemon] = v
-		as[ann.VersionCheck] = ann.Required
-	}
-	if as[ann.RootDaemon] == ann.Required {
-		if err = EnsureRootDaemonRunning(ctx); err != nil {
-			return err
-		}
-	}
-	if v := as[ann.UserDaemon]; v == ann.Optional || v == ann.Required {
-		if ctx, err = ensureUserDaemon(ctx, v == ann.Required); err != nil {
-			if v == ann.Optional && err == ErrNoUserDaemon {
-				// This is OK, but further initialization is not possible
-				err = nil
-			}
-			return err
-		}
-
-		// RootDaemon == Optional means that the RootDaemon must be started if
-		// the UserDaemon was started
-		if as[ann.RootDaemon] == ann.Optional {
-			if err = EnsureRootDaemonRunning(ctx); err != nil {
-				return err
-			}
-		}
-	} else {
-		// The rest requires a user daemon
-		return nil
-	}
-	if as[ann.VersionCheck] == ann.Required {
-		if err = ensureDaemonVersion(ctx); err != nil {
-			return err
-		}
-	}
-
-	if v := as[ann.Session]; v == ann.Optional || v == ann.Required {
-		if ctx, err = ensureSession(ctx, v == ann.Required); err != nil {
-			return err
-		}
-	}
-	cmd.SetContext(ctx)
 	return nil
 }