+++ pkg/client/k8s_config.go
@@ -1,4 +1,4 @@
-package k8s
+package client
 
 import (
 	"context"
@@ -7,22 +7,20 @@ import (
 	"os"
 
 	"github.com/spf13/pflag"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"gopkg.in/yaml.v3"
+	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
-	_ "k8s.io/client-go/plugin/pkg/client/auth" // Important for various cloud provider auth
 	"k8s.io/client-go/rest"
 
 	"github.com/datawire/dlib/dlog"
-
-	"github.com/telepresenceio/telepresence/v2/pkg/client"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/errcat"
 	"github.com/telepresenceio/telepresence/v2/pkg/iputil"
 	"github.com/telepresenceio/telepresence/v2/pkg/maps"
 )
 
-// The dnsConfig is part of the kubeconfigExtension struct.
-type dnsConfig struct {
+// The DnsConfig is part of the KubeconfigExtension struct.
+type DnsConfig struct {
 	// LocalIP is the address of the local DNS server. This entry is only
 	// used on Linux system that are not configured to use systemd-resolved and
 	// can be overridden by using the option --dns on the command line and defaults
@@ -42,25 +40,25 @@ type dnsConfig struct {
 	IncludeSuffixes []string `json:"include-suffixes,omitempty"`
 
 	// The maximum time to wait for a cluster side host lookup.
-	LookupTimeout metav1.Duration `json:"lookup-timeout,omitempty"`
+	LookupTimeout v1.Duration `json:"lookup-timeout,omitempty"`
 }
 
-// The managerConfig is part of the kubeconfigExtension struct. It configures discovery of the traffic manager.
-type managerConfig struct {
+// The ManagerConfig is part of the KubeconfigExtension struct. It configures discovery of the traffic manager.
+type ManagerConfig struct {
 	// Namespace is the name of the namespace where the traffic manager is to be found
 	Namespace string `json:"namespace,omitempty"`
 }
 
-// kubeconfigExtension is an extension read from the selected kubeconfig Cluster.
-type kubeconfigExtension struct {
-	DNS        *dnsConfig       `json:"dns,omitempty"`
+// KubeconfigExtension is an extension read from the selected kubeconfig Cluster.
+type KubeconfigExtension struct {
+	DNS        *DnsConfig       `json:"dns,omitempty"`
 	AlsoProxy  []*iputil.Subnet `json:"also-proxy,omitempty"`
 	NeverProxy []*iputil.Subnet `json:"never-proxy,omitempty"`
-	Manager    *managerConfig   `json:"manager,omitempty"`
+	Manager    *ManagerConfig   `json:"manager,omitempty"`
 }
 
-type Config struct {
-	kubeconfigExtension
+type Kubeconfig struct {
+	KubeconfigExtension
 	Namespace   string // default cluster namespace.
 	Context     string
 	Server      string
@@ -71,7 +69,7 @@ type Config struct {
 
 const configExtension = "telepresence.io"
 
-func NewConfig(c context.Context, flagMap map[string]string) (*Config, error) {
+func NewKubeconfig(c context.Context, flagMap map[string]string) (*Kubeconfig, error) {
 	// Namespace option will be passed only when explicitly needed. The k8Cluster is namespace agnostic with
 	// respect to this option.
 	delete(flagMap, "namespace")
@@ -135,7 +133,7 @@ func NewConfig(c context.Context, flagMap map[string]string) (*Config, error) {
 		namespace = "default"
 	}
 
-	k := &Config{
+	k := &Kubeconfig{
 		Context:     ctxName,
 		Server:      cluster.Server,
 		Namespace:   namespace,
@@ -145,24 +143,24 @@ func NewConfig(c context.Context, flagMap map[string]string) (*Config, error) {
 	}
 
 	if ext, ok := cluster.Extensions[configExtension].(*runtime.Unknown); ok {
-		if err = json.Unmarshal(ext.Raw, &k.kubeconfigExtension); err != nil {
+		if err = json.Unmarshal(ext.Raw, &k.KubeconfigExtension); err != nil {
 			return nil, errcat.Config.Newf("unable to parse extension %s in kubeconfig: %w", configExtension, err)
 		}
 	}
 
-	if k.kubeconfigExtension.Manager == nil {
-		k.kubeconfigExtension.Manager = &managerConfig{}
+	if k.KubeconfigExtension.Manager == nil {
+		k.KubeconfigExtension.Manager = &ManagerConfig{}
 	}
 
-	if k.kubeconfigExtension.Manager.Namespace == "" {
-		k.kubeconfigExtension.Manager.Namespace = client.GetEnv(c).ManagerNamespace
+	if k.KubeconfigExtension.Manager.Namespace == "" {
+		k.KubeconfigExtension.Manager.Namespace = GetEnv(c).ManagerNamespace
 	}
 
 	return k, nil
 }
 
 // This represents an inClusterConfig.
-func NewInClusterConfig(c context.Context, flagMap map[string]string) (*Config, error) {
+func NewInClusterConfig(c context.Context, flagMap map[string]string) (*Kubeconfig, error) {
 	// Namespace option will be passed only when explicitly needed. The k8Cluster is namespace agnostic with
 	// respect to this option.
 	delete(flagMap, "namespace")
@@ -187,16 +185,16 @@ func NewInClusterConfig(c context.Context, flagMap map[string]string) (*Config, 
 		namespace = "default"
 	}
 
-	return &Config{
+	return &Kubeconfig{
 		Namespace:   namespace,
 		Server:      restConfig.Host,
 		flagMap:     flagMap,
 		ConfigFlags: configFlags,
 		RestConfig:  restConfig,
 		// it may be empty, but we should avoid nil deref
-		kubeconfigExtension: kubeconfigExtension{
-			Manager: &managerConfig{
-				Namespace: client.GetEnv(c).ManagerNamespace,
+		KubeconfigExtension: KubeconfigExtension{
+			Manager: &ManagerConfig{
+				Namespace: GetEnv(c).ManagerNamespace,
 			},
 		},
 	}, nil
@@ -204,43 +202,49 @@ func NewInClusterConfig(c context.Context, flagMap map[string]string) (*Config, 
 
 // ContextServiceAndFlagsEqual determines if this instance is equal to the given instance with respect to context,
 // server, and flag arguments.
-func (kf *Config) ContextServiceAndFlagsEqual(okf *Config) bool {
+func (kf *Kubeconfig) ContextServiceAndFlagsEqual(okf *Kubeconfig) bool {
 	return kf != nil && okf != nil &&
 		kf.Context == okf.Context &&
 		kf.Server == okf.Server &&
 		maps.Equal(kf.flagMap, okf.flagMap)
 }
 
-func (kf *Config) GetManagerNamespace() string {
-	return kf.kubeconfigExtension.Manager.Namespace
+func (kf *Kubeconfig) GetManagerNamespace() string {
+	return kf.KubeconfigExtension.Manager.Namespace
 }
 
-func (kf *Config) GetRestConfig() *rest.Config {
+func (kf *Kubeconfig) GetRestConfig() *rest.Config {
 	return kf.RestConfig
 }
 
-func (kf *Config) AddRemoteKubeConfigExtension(ctx context.Context, cfgJson string) error {
-	dlog.Debugf(ctx, "Applying remote kubeconfig: %s", cfgJson)
-	remote := &kubeconfigExtension{}
-	if err := json.Unmarshal([]byte(cfgJson), &remote); err != nil {
+func (kf *Kubeconfig) AddRemoteKubeConfigExtension(ctx context.Context, cfgYaml []byte) error {
+	dlog.Debugf(ctx, "Applying remote dns and routing: %s", cfgYaml)
+	remote := struct {
+		DNS     *DNS     `yaml:"dns,omitempty"`
+		Routing *Routing `yaml:"routing,omitempty"`
+	}{}
+	if err := yaml.Unmarshal(cfgYaml, &remote); err != nil {
 		return fmt.Errorf("unable to parse remote kubeconfig: %w", err)
 	}
 	if kf.DNS == nil {
-		kf.DNS = remote.DNS
-	} else {
+		kf.DNS = &DnsConfig{}
+	}
+	if dns := remote.DNS; dns != nil {
 		if kf.DNS.LocalIP == "" {
-			kf.DNS.LocalIP = remote.DNS.LocalIP
+			kf.DNS.LocalIP = iputil.IPKey(dns.LocalIP)
 		}
 		if kf.DNS.RemoteIP == "" {
-			kf.DNS.RemoteIP = remote.DNS.RemoteIP
+			kf.DNS.RemoteIP = iputil.IPKey(dns.RemoteIP)
 		}
-		kf.DNS.ExcludeSuffixes = append(kf.DNS.ExcludeSuffixes, remote.DNS.ExcludeSuffixes...)
-		kf.DNS.IncludeSuffixes = append(kf.DNS.IncludeSuffixes, remote.DNS.IncludeSuffixes...)
+		kf.DNS.ExcludeSuffixes = append(kf.DNS.ExcludeSuffixes, dns.ExcludeSuffixes...)
+		kf.DNS.IncludeSuffixes = append(kf.DNS.IncludeSuffixes, dns.IncludeSuffixes...)
 		if kf.DNS.LookupTimeout.Duration == 0 {
-			kf.DNS.LookupTimeout = remote.DNS.LookupTimeout
+			kf.DNS.LookupTimeout.Duration = dns.LookupTimeout
 		}
 	}
-	kf.AlsoProxy = append(kf.AlsoProxy, remote.AlsoProxy...)
-	kf.NeverProxy = append(kf.NeverProxy, remote.NeverProxy...)
+	if routing := remote.Routing; routing != nil {
+		kf.AlsoProxy = append(kf.AlsoProxy, routing.AlsoProxy...)
+		kf.NeverProxy = append(kf.NeverProxy, routing.NeverProxy...)
+	}
 	return nil
 }