+++ pkg/client/rootd/session.go
@@ -222,13 +222,12 @@ func NewSession(c context.Context, scout *scout.Reporter, mi *rpc.OutboundInfo) 
 	if err != nil {
 		dlog.Warnf(c, "Failed to get remote config from traffic manager: %v", err)
 	} else {
-		bs := []byte(cliCfg.ConfigJson)
-		err := yaml.Unmarshal(bs, &cfg)
+		err := yaml.Unmarshal(cliCfg.ConfigYaml, &cfg)
 		if err != nil {
 			dlog.Warnf(c, "Failed to deserialize remote config: %v", err)
 		}
 	}
-	dlog.Debugf(c, "Traffic manager gave config %s", cliCfg.ConfigJson)
+	dlog.Debugf(c, "Traffic manager gave config %s", string(cliCfg.ConfigYaml))
 
 	as := convertSubnets(mi.AlsoProxySubnets)
 	ns := convertSubnets(mi.NeverProxySubnets)
@@ -312,10 +311,13 @@ func (s *Session) legacyClusterLookup(ctx context.Context, q *dns2.Question) (rr
 	return rrs, dns2.RcodeSuccess, nil
 }
 
-func (s *Session) getInfo() *rpc.OutboundInfo {
+func (s *Session) getNetworkConfig() *rpc.NetworkConfig {
 	info := rpc.OutboundInfo{
 		Session: s.session,
 		Dns:     s.dnsServer.GetConfig(),
+	}
+	nc := &rpc.NetworkConfig{
+		OutboundInfo: &info,
 	}
 	if s.dnsLocalAddr != nil {
 		info.Dns.RemoteIp = s.dnsLocalAddr.IP
@@ -333,8 +335,11 @@ func (s *Session) getInfo() *rpc.OutboundInfo {
 			info.NeverProxySubnets[i] = iputil.IPNetToRPC(np.RoutedNet)
 		}
 	}
-
-	return &info
+	nc.Subnets = make([]*manager.IPNet, len(s.curSubnets))
+	for i, sn := range s.curSubnets {
+		nc.Subnets[i] = iputil.IPNetToRPC(sn)
+	}
+	return nc
 }
 
 func (s *Session) configureDNS(dnsIP net.IP, dnsLocalAddr *net.UDPAddr) {